+++ packages/sui-test/CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+# 8.33.0 (2024-02-01)
+
+
+### Features
+
+* **packages/sui-test:** update alias path resolve ([d7ccb29](https://github.com/SUI-Components/sui/commit/d7ccb2957dff5ee9b18a23fac298d85261cda9d7))
+* **packages/sui-test:** update react aliases ([d051349](https://github.com/SUI-Components/sui/commit/d051349bf2be49064c881073de2def20c8f7d132))
+
+
+
 # 8.32.0 (2024-01-29)